Move `DnsQuery` into `dns` module
